requirements.txt
google-generativeai==0.7.2
pillow==10.4.0
python-dotenv==1.0.1

# ---- Core framework ----
fastapi==0.111.0
uvicorn[standard]==0.30.1

# ---- Async database + Postgres ----
asyncpg==0.29.0
psycopg2-binary==2.9.9       # optional if someone uses sync code or pgvector CLI setup
sqlalchemy==2.0.31           # ORM for database models
supabase==2.3.4              # Supabase Python client
vecs==0.4.0                  # Supabase vector client for pgvector

# ---- HTTP requests (used for APIs: OpenAI, Cohere, etc.) ----
httpx==0.27.0

# ---- Data handling & utils ----
pydantic==2.8.2
python-dotenv==1.0.1

# ---- Embeddings / Rerank / LLM APIs ----
openai==1.45.0               # for embeddings + chat completions
cohere==5.5.8                # for reranking (optional)
# voyageai==0.2.3            # uncomment if you switch provider
# jinaai==0.3.9              # uncomment if using Jina embeddings

# ---- Vector math (optional helpers for local checks) ----
numpy==1.26.4
scikit-learn==1.5.2

# ---- (Optional) Dev & testing ----
black==24.4.2
pytest==8.3.2

psycopg2-binary
pgvector
